requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
sqlalchemy
psycopg2-binary
alembic
pgvector
fastapi
uvicorn
python-multipart
